address comments
- swap parameter for test case asserts
- remove excessive null check (create noop impl for rate limiter)
